AnnotatedValue impl split into several files
